+++ packages/router/test/adapters/subgraph/subgraph.spec.ts
@@ -31,6 +31,7 @@ type SdkMock = {
   GetTransaction: SinonStub;
   GetAssetBalance: SinonStub;
   GetBlockNumber: SinonStub;
+  GetTransactionsWithRouter: SinonStub;
 };
 
 let sdks: Record<number, SinonStubbedInstance<FallbackSubgraph<SdkMock>>>;
@@ -62,6 +63,7 @@ describe("Subgraph Adapter", () => {
       GetTransaction: stub().resolves(undefined),
       GetAssetBalance: stub().resolves(constants.Zero),
       GetBlockNumber: stub().resolves({ _meta: { block: { number: 10000 } } }),
+      GetTransactionsWithRouter: stub().resolves({ transactions: [] }),
     };
 
     txServiceMock.getBlockNumber.resolves(10000);
@@ -188,7 +190,7 @@ describe("Subgraph Adapter", () => {
         },
       });
       const testError = new Error("fail");
-      sdk.GetTransactions.rejects(testError);
+      sdk.GetTransactionsWithRouter.rejects(testError);
 
       try {
         await getActiveTransactions(requestContextMock);
@@ -249,7 +251,7 @@ describe("Subgraph Adapter", () => {
           },
         };
       });
-      sdk.GetTransactions.onCall(0).callsFake(async ({ transactionIds }) => {
+      sdk.GetTransactionsWithRouter.onCall(0).callsFake(async ({ transactionIds }) => {
         const expectedId = transactionSubgraphMock.transactionId.toLowerCase();
         expect(transactionIds).to.deep.eq([expectedId]);
         return {
@@ -280,7 +282,7 @@ describe("Subgraph Adapter", () => {
           },
         };
       });
-      sdk.GetTransactions.onCall(0).callsFake(async ({ transactionIds }) => {
+      sdk.GetTransactionsWithRouter.onCall(0).callsFake(async ({ transactionIds }) => {
         const expectedId = transactionSubgraphMock.transactionId.toLowerCase();
         expect(transactionIds).to.deep.eq([expectedId]);
         return {
@@ -310,7 +312,7 @@ describe("Subgraph Adapter", () => {
           },
         };
       });
-      sdk.GetTransactions.onCall(0).callsFake(async ({ transactionIds }) => {
+      sdk.GetTransactionsWithRouter.onCall(0).callsFake(async ({ transactionIds }) => {
         const expectedId = transactionSubgraphMock.transactionId.toLowerCase();
         expect(transactionIds).to.deep.eq([expectedId]);
         return {